model/training_data_collection/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask==3.1.0
psycopg==3.2.4
pymongo==4.11.1
python-dotenv==1.0.1
requests==2.32.3
gunicorn==23.0.0
